Add separate TranslationProvider
