+++ src/app/[locale]/layout.tsx
@@ -29,14 +29,14 @@ export default async function RootLayout({
   params: { locale: string }
   children: React.ReactNode
 }) {
-  const { t, resources } = await initTranslations(locale, i18nNamespaces);
+  const { i18n } = await initTranslations(locale, i18nNamespaces);
 
   return (
-    <html lang="en" className="h-full antialiased" suppressHydrationWarning>
+    <html lang={i18n.language} className="h-full antialiased" suppressHydrationWarning>
       <body className="flex h-full bg-zinc-50 dark:bg-black">
-        <Providers namespaces={i18nNamespaces} resources={resources} locale={locale}>
+        <Providers>
           <div className="flex w-full">
-            <Layout t={t}>{children}</Layout>
+            <Layout locale={locale}>{children}</Layout>
           </div>
         </Providers>
       </body>
