+++ src/app/[locale]/providers.tsx
@@ -2,9 +2,6 @@
 
 import React, {useEffect} from 'react'
 import {ThemeProvider, useTheme} from 'next-themes'
-import {createInstance, Resource} from "i18next";
-import initTranslations from "@/app/i18n";
-import {I18nextProvider} from "react-i18next";
 
 function ThemeWatcher() {
   let { resolvedTheme, setTheme } = useTheme()
@@ -30,24 +27,11 @@ function ThemeWatcher() {
   return null
 }
 
-export function Providers({ children, locale, namespaces, resources }:
-{
-  locale: string;
-  namespaces: string[];
-  resources: Resource
-  children: React.ReactNode
-}) {
-  const i18n = createInstance();
-
-  // noinspection JSIgnoredPromiseFromCall
-  initTranslations(locale, namespaces, i18n, resources);
-
+export function Providers({ children }: { children: React.ReactNode }) {
   return (
-    <I18nextProvider i18n={i18n}>
-      <ThemeProvider attribute="class" disableTransitionOnChange>
-        <ThemeWatcher />
-        {children}
-      </ThemeProvider>
-    </I18nextProvider>
+    <ThemeProvider attribute="class" disableTransitionOnChange>
+      <ThemeWatcher />
+      {children}
+    </ThemeProvider>
   )
 }
