+++ src/components/header/HeaderNavigation.tsx
@@ -11,7 +11,7 @@ export function DesktopNavigation(props: React.ComponentPropsWithoutRef<'nav'>) 
 
   return (
     <nav {...props}>
-      <ul className="flex rounded-full bg-white/90 px-3 text-sm font-medium text-zinc-800 shadow-lg shadow-zinc-800/5 ring-1 ring-zinc-900/5 backdrop-blur dark:bg-zinc-800/90 dark:text-zinc-200 dark:ring-white/10">
+      <ul className="flex rounded-full bg-white/90 px-3 whitespace-nowrap text-sm font-medium text-zinc-800 shadow-lg shadow-zinc-800/5 ring-1 ring-zinc-900/5 backdrop-blur dark:bg-zinc-800/90 dark:text-zinc-200 dark:ring-white/10">
         <NavItem href="/about">{t('navigation.about')}</NavItem>
         <NavItem href="/projects">{t('navigation.projects')}</NavItem>
       </ul>
